Create Account page wireframe.png
Here is the wireframe image for the account page.

On the account page you will be able to post the items you would like to sell to WeBuyYourGames.com.

Below the forms to post the information there will be your previous order details or the previous submission the user created.